backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
pwdlib[argon2]
PyJWT
pydantic
psycopg2-binary
